refactor: consolidate duplicate BridgeTransaction/BridgeTransactionData types (#248)
Move the canonical BridgeTransactionData definition (the more complete one,
with hash + memo) from stellar.ts to types.ts as the single source of truth.
Remove the now-redundant BridgeTransaction interface from types.ts.

- src/lib/types.ts: replace BridgeTransaction with BridgeTransactionData
  (adds memo field, otherwise identical fields)
- src/lib/stellar.ts: import BridgeTransactionData from types.ts; remove
  the local interface definition
- src/components/transaction-history.tsx: import BridgeTransactionData
  from @/lib/types instead of @/lib/stellar
- src/components/routes/dashboard-page.tsx: import BridgeTransactionData
  from @/lib/types; remove the BridgeTransaction alias

npm run typecheck: 0 errors in the four changed files (pre-existing
test errors in __tests__/ are unrelated to this change).
