+++ src/components/transaction-history.tsx
@@ -1,6 +1,6 @@
 import React, { memo, useMemo } from "react";
 import { ArrowLeftRight, CreditCard, Building2, ExternalLink, Loader2 } from "lucide-react";
-import type { BridgeTransactionData } from "@/lib/stellar";
+import type { BridgeTransactionData } from "@/lib/types";
 import { getExplorerUrl } from "@/lib/stellar";
 import type { StellarNetwork } from "@/lib/types";
